Split string difference helpers into focused files

Copilot-Session: ab5160cf-e6b4-4ba6-b58e-e3b8ba116305
